agentic_docs_squad/requirements.txt
google-adk
google-generativeai
python-dotenv
fastapi
uvicorn